+++ test-published-types/index.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const react = require('eslint-plugin-react');
+
+/** @type {import('eslint').Linter.Config[]} */
+module.exports = [
+  {
+    plugins: {
+      react,
+    },
+  },
+];